+++ jason/config/postgres.py
@@ -1,5 +1,3 @@
-import sqlalchemy.orm
-
 from jason.core.configuration import props
 
 
@@ -10,30 +8,3 @@ class PostgresConfigMixin:
     DB_USER = props.String(nullable=True)
     DB_PASS = props.String(nullable=True)
     TEST_DB_URL = props.String(default="sqlite:///:memory:")
-
-
-def scoped_session(factory):
-    """
-    creates a thread safe session factory
-
-    """
-    return sqlalchemy.orm.scoped_session(factory)
-
-
-def postgres_engine(config: PostgresConfigMixin, testing=False):
-    """
-    builds a connection uri based on config options and returns an engine.
-
-    NOTE: if testing is true, the uri will be to an in-memory sqlite database
-
-    """
-    if not testing:
-        if config.DB_USER:
-            credentials = f"{config.DB_USER}:{config.DB_PASS}@"
-        else:
-            credentials = ""
-        host = f"{config.DB_HOST}:{config.DB_PORT}"
-        db_url = f"{config.DB_DRIVER}://{credentials}{host}"
-    else:
-        db_url = config.TEST_DB_URL
-    return sqlalchemy.create_engine(db_url, echo=True)